Propose a simple docker image
Wrap the tool inside the docker to let its simple use without the need to install node in the environment. It could sound useless for some folks, but imagine if you need to handle a ruby, java, etc tools? Some will prefer to keep the environment clean and easy to backup. Docker will give this easy solution.